feat: redo the filter bar on mobile
