--- src/components/pages/templates/dropdown-filter-bar/index.js
@@ -1,3 +1,0 @@
-import DropdownFilterBar from './dropdown-filter-bar';
-
-export default DropdownFilterBar;
