+++ src/components/pages/templates/filter-bar/filter-bar.jsx
@@ -1,8 +1,31 @@
+import clsx from 'clsx';
+import { AnimatePresence, domAnimation, LazyMotion, m } from 'framer-motion';
 import PropTypes from 'prop-types';
 import { useEffect, useState } from 'react';
 
+import ChevronRightIcon from './images/chevron-right.inline.svg';
+
 const FilterGroup = ({ type, items, filteredTemplates, handleFilterChange }) => {
   const [filterCount, setFilterCount] = useState(0);
+  const [isOpen, setIsOpen] = useState(false);
+
+  const variants = {
+    open: {
+      opacity: 1,
+      height: 'auto',
+      transition: {
+        duration: 0.2,
+      },
+    },
+    closed: {
+      opacity: 0,
+      height: 0,
+    },
+  };
+
+  const toggleDropdown = () => {
+    setIsOpen(!isOpen);
+  };
 
   useEffect(() => {
     const count = filteredTemplates.filters[type.toLowerCase()]?.length || 0;
@@ -10,24 +33,48 @@ const FilterGroup = ({ type, items, filteredTemplates, handleFilterChange }) => 
   }, [filteredTemplates.filters, type]);
 
   return (
-    <li className="flex flex-col pb-5 pt-5 first:pt-0 last:pb-0" key={type}>
+    <li
+      className="flex flex-col border-b border-gray-new-80/80 pb-5 pt-5 first:pt-0 last:border-b-0 last:pb-0 dark:border-gray-new-15/80 lg:pb-4 lg:pt-4 lg:last:border-b last:lg:pb-4"
+      key={type}
+    >
       <fieldset>
-        <div className="flex items-center justify-between">
-          <legend className="block font-medium leading-tight tracking-extra-tight">{type}</legend>
+        <button
+          className="flex w-full items-center justify-between"
+          type="button"
+          onClick={toggleDropdown}
+        >
+          <div className="flex items-center gap-3">
+            <ChevronRightIcon
+              className={clsx('h-5 w-5 transition-transform duration-200', isOpen && 'rotate-90')}
+            />
+            <legend className="block font-medium leading-tight tracking-extra-tight lg:text-sm">
+              {type}
+            </legend>
+          </div>
           {filterCount > 0 && (
             <span className="inline-flex h-5 w-6 items-center justify-center rounded-full bg-gray-new-90 text-[10px] leading-none tracking-extra-tight dark:bg-gray-new-15">
               {filterCount}
             </span>
           )}
-        </div>
-        <div className="mt-4 flex flex-col gap-y-3">
-          {items.map(({ name, value }) => (
-            <label
-              className="flex cursor-pointer items-center gap-2.5 text-sm leading-tight text-gray-new-10 dark:text-gray-new-90"
-              key={value}
-            >
-              <input
-                className={`
+        </button>
+        <LazyMotion features={domAnimation}>
+          <AnimatePresence>
+            {isOpen && (
+              <m.div
+                initial="closed"
+                animate="open"
+                exit="closed"
+                variants={variants}
+                onMouseDown={(e) => e.preventDefault()}
+              >
+                <div className="flex flex-col gap-y-3 pt-4">
+                  {items.map(({ name, value }) => (
+                    <label
+                      className="flex cursor-pointer items-center gap-2.5 text-sm leading-tight text-gray-new-10 dark:text-gray-new-90"
+                      key={value}
+                    >
+                      <input
+                        className={`
                 h-4 w-4 appearance-none rounded-sm border
                 border-gray-new-70 bg-center bg-no-repeat
                 transition-colors duration-100
@@ -36,18 +83,23 @@ const FilterGroup = ({ type, items, filteredTemplates, handleFilterChange }) => 
                 dark:border-gray-new-30 dark:checked:border-white dark:checked:bg-white
                 dark:checked:bg-[url('/images/templates/check-dark-mode.svg')]
               `}
-                type="checkbox"
-                name={value}
-                checked={
-                  filteredTemplates.filters[type.toLowerCase()]?.includes(value.toLowerCase()) ||
-                  false
-                }
-                onChange={handleFilterChange(type, value)}
-              />
-              {name}
-            </label>
-          ))}
-        </div>
+                        type="checkbox"
+                        name={value}
+                        checked={
+                          filteredTemplates.filters[type.toLowerCase()]?.includes(
+                            value.toLowerCase()
+                          ) || false
+                        }
+                        onChange={handleFilterChange(type, value)}
+                      />
+                      {name}
+                    </label>
+                  ))}
+                </div>
+              </m.div>
+            )}
+          </AnimatePresence>
+        </LazyMotion>
       </fieldset>
     </li>
   );
@@ -73,7 +125,7 @@ const FilterBar = ({
 }) => (
   <form className={className}>
     <input
-      className="h-9 rounded border border-gray-new-80/80 bg-transparent bg-[url('/images/templates/search-light-mode.svg')] bg-[left_0.625rem_center] bg-no-repeat p-2.5 pl-[34px] text-sm leading-none tracking-extra-tight text-gray-new-70 focus:outline-none dark:border-gray-new-20 dark:bg-[url('/images/templates/search-dark-mode.svg')]"
+      className="h-9 rounded border border-gray-new-80/80 bg-transparent bg-[url('/images/templates/search-light-mode.svg')] bg-[left_0.625rem_center] bg-no-repeat p-2.5 pl-[34px] text-sm leading-none tracking-extra-tight text-gray-new-70 focus:outline-none dark:border-gray-new-20 dark:bg-[url('/images/templates/search-dark-mode.svg')] lg:w-full"
       type="search"
       placeholder="Search"
       onChange={handleSearch}
@@ -90,7 +142,7 @@ const FilterBar = ({
         </button>
       )}
     </div>
-    <ul className="mt-5 flex flex-col divide-y divide-gray-new-80/80 dark:divide-gray-new-15/80">
+    <ul className="mt-5 flex flex-col lg:mt-4">
       {filters.map(({ type, items }) => (
         <FilterGroup
           key={type}
